Changeing Migrations in that way, that it ist possible to migrate to version 0 and current. + Small fixes at re_queries

+++ db/migrate/011_create_re_attachments.rb
@@ -9,6 +9,9 @@ class CreateReAttachments < ActiveRecord::Migration
   end
 
   def self.down
-    drop_table :re_attachments
+    if table_exists?(:re_attachments)
+      drop_table :re_attachments 
+    end
   end
+  
 end
